fixed testing
