+++ src/model/game_pieces/dev_card.py
@@ -1,6 +1,9 @@
 from ..interfaces.i_dev_card import IDevCard
 from ..interfaces.i_item import IItem
 from ..interfaces.i_encounter import IEncounter
+from ..item import get_item
+from src.enums_and_types import ItemName
+from ..encounters.encounters import MessageEncounter
 
 
 class DevCard(IDevCard):
@@ -19,3 +22,16 @@ class DevCard(IDevCard):
         assert 9 <= time <= 11, \
             f"Cannot get an ecounter from dev card at time {time}pm"
         return self._encounters[time - 9]
+
+    @staticmethod
+    def get_dev_cards() -> list[IDevCard]:
+        return [
+            DevCard(
+                get_item(ItemName.OIL),
+                MessageEncounter()
+            )
+        ]
+
+        def create_encounter(encounter: IEncounter, values) -> IEncounter:
+            encounter.set_values(values)
+            return encounter
