+++ src/model/game_pieces/tile.py
@@ -1,4 +1,4 @@
-from enums_and_types import *
+from src.enums_and_types import *
 from ..interfaces.i_tile import ITile
 from ..interfaces.i_encounter import IEncounter
