+++ src/model/item/weapon.py
@@ -1,4 +1,4 @@
-from enums_and_types import ItemName, ItemType
+from src.enums_and_types import ItemName, ItemType
 from .item import Item
