+++ test/game_over/test_game_over.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
-from ...interfaces import IGameOver
-from ..game_over import GameOver
-from enums_and_types import GameOverReason
+from src.model.interfaces import IGameOver
+from src.model.game_over import GameOver
+from src.enums_and_types import GameOverReason
 
 
 class TestGameOver(TestCase):
