+++ test/item/test_candle.py
@@ -1,18 +1,18 @@
 from unittest import TestCase
-from ..item_helper import IItem, get_item
-from enums_and_types import ItemName, ItemType
+from src.model.item.item_helper import IItem, get_item
+from src.enums_and_types import ItemName, ItemType
 
 
-class TestGasoline(TestCase):
+class TestCandle(TestCase):
 
     def setUp(self):
-        self._item: IItem = get_item(ItemName.GASOLINE)
+        self._item: IItem = get_item(ItemName.CANDLE)
 
-    def test_name_is_gasoline(self):
-        expected = ItemName.GASOLINE
+    def test_name_is_candle(self):
+        expected = ItemName.CANDLE
         actual = self._item.name
         self.assertEqual(expected, actual)
-    
+
     def test_type_is_combine_only(self):
         expected = ItemType.COMBINE_ONLY
         actual = self._item.type
@@ -29,8 +29,8 @@ class TestGasoline(TestCase):
         actual = self._item.uses_remaining
         self.assertEqual(expected, actual)
 
-    def test_combinable_with_candle(self):
-        self.assertIn(ItemName.CANDLE, self._item.combinable_with)
+    def test_combinable_with_oil(self):
+        self.assertIn(ItemName.OIL, self._item.combinable_with)
 
-    def test_combinable_with_chainsaw(self):
-        self.assertIn(ItemName.CHAINSAW, self._item.combinable_with)
+    def test_combinable_with_gasoline(self):
+        self.assertIn(ItemName.GASOLINE, self._item.combinable_with)
