+++ test/item/test_grisly_femur.py
@@ -1,15 +1,15 @@
 from unittest import TestCase
-from ..item_helper import IItem, get_item
-from enums_and_types import ItemName, ItemType
+from src.model.item.item_helper import IItem, get_item
+from src.enums_and_types import ItemName, ItemType
 
 
-class TestMachete(TestCase):
+class TestGrislyFemur(TestCase):
 
     def setUp(self):
-        self._item: IItem = get_item(ItemName.MACHETE)
+        self._item: IItem = get_item(ItemName.GRISLY_FEMUR)
 
-    def test_name_is_machete(self):
-        expected = ItemName.MACHETE
+    def test_name_is_grisly_femur(self):
+        expected = ItemName.GRISLY_FEMUR
         actual = self._item.name
         self.assertEqual(expected, actual)
 
@@ -30,7 +30,7 @@ class TestMachete(TestCase):
         actual = self._item.combinable_with
         self.assertEqual(expected, actual)
 
-    def test_attack_bonus_is_two(self):
-        expected = 2
+    def test_attack_bonus_is_one(self):
+        expected = 1
         actual = self._item.attack_bonus
         self.assertEqual(expected, actual)
