+++ test/item/test_machete.py
@@ -1,15 +1,15 @@
 from unittest import TestCase
-from ..item_helper import IItem, get_item
-from enums_and_types import ItemName, ItemType
+from src.model.item.item_helper import IItem, get_item
+from src.enums_and_types import ItemName, ItemType
 
 
-class TestGrislyFemur(TestCase):
+class TestMachete(TestCase):
 
     def setUp(self):
-        self._item: IItem = get_item(ItemName.GOLF_CLUB)
+        self._item: IItem = get_item(ItemName.MACHETE)
 
-    def test_name_is_golf_club(self):
-        expected = ItemName.GOLF_CLUB
+    def test_name_is_machete(self):
+        expected = ItemName.MACHETE
         actual = self._item.name
         self.assertEqual(expected, actual)
 
@@ -30,7 +30,7 @@ class TestGrislyFemur(TestCase):
         actual = self._item.combinable_with
         self.assertEqual(expected, actual)
 
-    def test_attack_bonus_is_one(self):
-        expected = 1
+    def test_attack_bonus_is_two(self):
+        expected = 2
         actual = self._item.attack_bonus
         self.assertEqual(expected, actual)
